Use PositiveBigIntegerField for node memory field

+++ analytics/analytics/migrations/0006_jobpod_node_alter_job_aws_alter_job_duration_and_more.py
@@ -71,7 +71,7 @@ class Migration(migrations.Migration):
                 ("name", models.CharField(max_length=64)),
                 ("system_uuid", models.UUIDField()),
                 ("cpu", models.PositiveIntegerField()),
-                ("memory", models.PositiveIntegerField()),
+                ("memory", models.PositiveBigIntegerField()),
                 (
                     "capacity_type",
                     models.CharField(
